GUIs for Main Menu and New Game
Implemented prototype gui for main menu and new game. Allows main menu to go to new game(char creator) but still needs to implement load game, etc. Also added ViewManager which is supposed to intialize all the views with their models and controllers (Might need to rename the class in that case).